// IsveskiDemo_NodeJS/clientcode3/model/baseAction.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ActionTypeEnum } from './actionTypeEnum';

export class BaseAction {
    'actionType': ActionTypeEnum;

    static discriminator: string | undefined = "actionType";

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "actionType",
            "baseName": "actionType",
            "type": "ActionTypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return BaseAction.attributeTypeMap;
    }
}

export namespace BaseAction {
}