// IsveskiDemo_NodeJS/clientcode3/model/bioAuthenticationRequest.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class BioAuthenticationRequest {
    'communicationId': string;
    'title': string;
    'allowAlternative': boolean;
    'reason': string;
    'timeoutSek': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "communicationId",
            "baseName": "communicationId",
            "type": "string"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "allowAlternative",
            "baseName": "allowAlternative",
            "type": "boolean"
        },
        {
            "name": "reason",
            "baseName": "reason",
            "type": "string"
        },
        {
            "name": "timeoutSek",
            "baseName": "timeoutSek",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return BioAuthenticationRequest.attributeTypeMap;
    }
}