// IsveskiDemo_NodeJS/clientcode3/model/bioMetricAuthenticationResponse.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { AuthenticationMethod } from './authenticationMethod';

export class BioMetricAuthenticationResponse {
    'succeeded': boolean;
    'authenticationMethod': AuthenticationMethod;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "succeeded",
            "baseName": "succeeded",
            "type": "boolean"
        },
        {
            "name": "authenticationMethod",
            "baseName": "authenticationMethod",
            "type": "AuthenticationMethod"
        }    ];

    static getAttributeTypeMap() {
        return BioMetricAuthenticationResponse.attributeTypeMap;
    }
}

export namespace BioMetricAuthenticationResponse {
}