// IsveskiDemo_NodeJS/clientcode3/model/createTicketLogRequest.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { CreateTicketLogRequestLogTemplate } from './createTicketLogRequestLogTemplate';

export class CreateTicketLogRequest {
    'ticketId': string;
    'logTemplate': CreateTicketLogRequestLogTemplate;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "ticketId",
            "baseName": "ticketId",
            "type": "string"
        },
        {
            "name": "logTemplate",
            "baseName": "logTemplate",
            "type": "CreateTicketLogRequestLogTemplate"
        }    ];

    static getAttributeTypeMap() {
        return CreateTicketLogRequest.attributeTypeMap;
    }
}