// IsveskiDemo_NodeJS/clientcode3/model/createTicketLogRequestLogTemplate.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { LogTemplate1 } from './logTemplate1';
import { LogTemplateTypeEnum } from './logTemplateTypeEnum';

export class CreateTicketLogRequestLogTemplate {

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
    ];

    static getAttributeTypeMap() {
        return CreateTicketLogRequestLogTemplate.attributeTypeMap;
    }
}

export namespace CreateTicketLogRequestLogTemplate {
}