// IsveskiDemo_NodeJS/clientcode3/model/createTicketRequestTemplate.ts
/**
 * Isveski Client
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Template } from './template';
import { Template1 } from './template1';
import { TemplateTypeEnum } from './templateTypeEnum';

export class CreateTicketRequestTemplate {
    'image'?: string | null;
    'title'?: string | null;
    'description'?: string | null;
    'time'?: Date | null;
    'expiry'?: Date | null;
    'templateType': TemplateTypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "image",
            "baseName": "image",
            "type": "string"
        },
        {
            "name": "title",
            "baseName": "title",
            "type": "string"
        },
        {
            "name": "description",
            "baseName": "description",
            "type": "string"
        },
        {
            "name": "time",
            "baseName": "time",
            "type": "Date"
        },
        {
            "name": "expiry",
            "baseName": "expiry",
            "type": "Date"
        },
        {
            "name": "templateType",
            "baseName": "templateType",
            "type": "TemplateTypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return CreateTicketRequestTemplate.attributeTypeMap;
    }
}

export namespace CreateTicketRequestTemplate {
}